In this commit i have finished the bottom navigation bar. Also i have created 4 empty fragments for quran , ahadeth , sebha , radio.
