+++ app/src/main/AndroidManifest.xml
@@ -12,21 +12,27 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Islami"
         tools:targetApi="31">
+
         <activity
-            android:name=".sebha.Sebha"
+            android:name=".fragments.fragments.sebha.Sebha"
             android:exported="false" />
+
         <activity
-            android:name=".home.HomeActivity"
-            android:exported="false" />
+            android:name=".fragments.fragments.HomeActivity"
+            android:exported="true">
+
+
+        </activity>
+
         <activity
             android:name=".splash.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
     </application>
 
 </manifest>
