--- app/src/main/java/com/example/islami/home/HomeActivity.kt
@@ -1,15 +1,0 @@
-package com.example.islami.home
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.islami.R
-import com.example.islami.databinding.ActivityHomeBinding
-
-class HomeActivity : AppCompatActivity() {
-    lateinit var binding: ActivityHomeBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityHomeBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-}
