+++ app/src/main/res/layout/activity_sebha.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".sebha.Sebha"
+    tools:context=".fragments.fragments.sebha.Sebha"
     android:background="@drawable/main_background"
     >
     <TextView
@@ -97,7 +97,9 @@
         android:layout_height="wrap_content"
         android:background="@color/orange"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+
+        />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
