+++ app/src/main/res/menu/main_bottom_nav_menu.xml
@@ -1,4 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
+<item
+    android:id="@+id/ic_quran"
+    android:title="@string/quran"
+    android:icon="@drawable/icon_quran"/>
 
+    <item
+        android:id="@+id/ic_ahadeth"
+        android:title="@string/ahadeth"
+        android:icon="@drawable/icon_hadeth"/>
+
+    <item
+        android:id="@+id/ic_sebha"
+        android:title="@string/sebha"
+        android:icon="@drawable/icon_sebha"/>
+
+    <item
+        android:id="@+id/ic_radio"
+        android:title="@string/radio"
+        android:icon="@drawable/icon_radio"/>
 </menu>
